+++ internal/client/storage/storage.go
@@ -8,7 +8,7 @@ import (
 
 type DB interface {
 	List(query model.ListQuery) (data []model.DBItem, err error)
-	Count(query model.ListQuery) (n int, err error)
+	Count(query model.ListQuery) (n uint64, err error)
 	Get(key string) (data model.DBRecord, err error)
 	Save(data model.DBRecord) (err error)
 	Delete(key string) (err error)
